feat: 弹幕字段类型调整

+++ business/src/main/java/com/abdecd/moebackend/business/pojo/vo/danmaku/DanmakuVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class DanmakuVO {
 //    Long id;
-    Integer begin;
+    Double begin;
     Integer mode;
     Integer size;
     String color;
